Add charge-pump for driving motor/pwm rail FET
As I'm using an n-channel FET to switch the motor/pwm rail a charge-
pump is required to obtain a high voltage for driving the gate.

This charge pump uses a dual op-amp as a relaxation oscillator and
unity gain buffer. The buffer drives a cap/diode network to give
~22V on the output.