Add 2 bit level translator to the BB UART and connect it to the MSP
The level translator is in an SSOP8 package but the gEDA SSOP8
footprint is way too wide so I've selected the TSSOP8 footprint, it
seems to fit the drawing much better. NEEDS TO BE CHECKED.

The TX and RX lines from the BB connect to their opposite counterparts
on the MSP. I.E. BB TX to MSP RX, BB RX to MSP TX.

The BB_TX and BB_RX lines are now on the 3.3V side of the translator.