Connect BB expansion header signals to MSP
Connected the beagleboard's reset line to the MSP via a FET to pull
it low. The REGEN line from the beagleboard is open-drain and can be
used to detect if the beagleboard is powered on. The other two lines
that we have access to on the beagleboard (that are not MMC/SD related)
have been tied to the MSP via some bi-directional level translation FETs.