Add a charging LED. Illuminates when the charger is plugged in
Unfortunately there is no way I can think of to have an LED light
up when the battery is charging and the power board is switched off.
The best we can have is an LED which lights up when the charger is
plugged in. If the power board is switched on then we can have
another LED show the charging status and display something on the
screen.