Revert "Add a charging LED. Illuminates when the charger is plugged in"
This reverts commit 69c8200de19814e232f60e632d0b3e12e14cdaa3.

It will be too misleading to have an LED that lights up when you plug the
charger in as it doesn't necessarily mean that the battery is charging.
The charger could be off, battery not plugged in, fuse blown, etc.

We'll just have to live with only showing the charging status when the
power board is switched on.