Rename the 'Logic Power Switch' to 'Power Switch' and change its footprint
The logic power switch is now the only switch used to control the power
to the kit, there is no more 'charge/run' switch. This means we can now
safely call it the 'power switch' without causing confusion.

It will be integrated into the power board casing assembly and will connect
to the power board via a 2 way header. The 2 way header is a right angle
type due to the restricted height in the power board case.