Add connector to 5V rail for a powered USB hub
I've measured the current drawn from a USB port though on of the
Belkin hubs with a webcam and memory stick plugged in. It comes to
around 350mA. When attempting to plug a keyboard in the hub cuts
power to one of the other ports so it needs to be powered to support
more than the normal webcam+memory stick.