Give board a name to allow screen capture script to ID the correct window
I'm going to record the process of laying out the power board to make a
cool video. A screenshot is taken of the PCB window every 2 seconds.
I've used ImageMagick's 'import' program which allows you to capture a
certain window by passing its name, setting the board name changes the
window title allowing it to be uniquely identified.