Change 5V SMPS to be enabled by default
At the moment it's not possible to reset the MSP without the beagleboard
resetting due to the 5V SMPS being off when the 5V enable line isn't being
driven. This changes the pull-up to a pull-down, causing the 5V SMPS to be
enabled rather than disabled by default.

To ensure that the 5V rail doesn't blip on when the board is first switched
on there is a cap between the 3.3V rail and the enable line which introduces
a delay of 10ms between the 3.3V rail coming up and the 5V SMPS being powered
up. In this time the MSP can pull the line high, stopping the SMPS turn on
until everything else has been checked.