Resize and move the silicone shim cutouts
After showing the result to a power board, it needs to look more like
this.  I made a couple of assumptions when initially inputting this.
One of them was that the shim goes to the edge of the power board's
case, when in fact it goes to the edge of the screen.  This commit
moves the holes and resizes them to account for this.

TO THE LASERS!