Add TVS diode and bulk decoupling cap to battery input (#698)
This has already been applied to existing power boards as the #698 mod,
but never made its way to the schematic. They're both after reverse polarity
protection FET.

22uF was chosen for the filtering cap since I tried a 2mF filtering cap
for testing #1260, to find a discharge time of a couple of minutes on the
12V rail.

Note that a lot of the changes in this commit are just re-jiggling other
components in the schematic to accomodate the two new components.